+++ include/BSMPT/baryo_calculation/transport_equations.h
@@ -1,21 +1,8 @@
-/*
- * transport_equations.h
- *
- *  Copyright (C) 2020  Philipp Basler, Margarete Mühlleitner and Jonas Müller
+// Copyright (C) 2020  Philipp Basler, Margarete Mühlleitner and Jonas Müller
+// SPDX-FileCopyrightText: 2021 Philipp Basler, Margarete Mühlleitner and Jonas Müller
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
 
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
 #ifndef SRC_transport_equations_H_
 #define SRC_transport_equations_H_