+++ include/BSMPT/minimizer/Minimizer.h
@@ -1,21 +1,7 @@
-/*
- * Minimizer.h
- *
- *  Copyright (C) 2018  Philipp Basler and Margarete Mühlleitner
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
+// Copyright (C) 2018  Philipp Basler and Margarete Mühlleitner
+// SPDX-FileCopyrightText: 2021 Philipp Basler, Margarete Mühlleitner and Jonas Müller
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
 
 /**
  * @file
@@ -114,7 +100,8 @@ Minimize_gen_all(const std::shared_ptr<Class_Potential_Origin> &modelPointer,
                  const double &Temp,
                  std::vector<double> &Check,
                  const std::vector<double> &start,
-                 const int &WhichMinimizer = WhichMinimizerDefault);
+                 const int &WhichMinimizer = WhichMinimizerDefault,
+                 bool UseMultithreading    = true);
 
 /**
  * @brief The MinimizerStatus enum for the Statusflags of the minimizer
@@ -172,7 +159,8 @@ EWPTReturnType
 PTFinder_gen_all(const std::shared_ptr<Class_Potential_Origin> &modelPointer,
                  const double &TempStart,
                  const double &TempEnd,
-                 const int &WhichMinimizer = WhichMinimizerDefault);
+                 const int &WhichMinimizer = WhichMinimizerDefault,
+                 bool UseMultithreading    = true);
 
 /**
  * @brief Minimize_gen_all_tree_level Minimizes the tree-level potential
@@ -191,7 +179,8 @@ Minimize_gen_all_tree_level(const ModelID::ModelIDs &Model,
                             const std::vector<double> &parCT,
                             std::vector<double> &Check,
                             const std::vector<double> &start,
-                            int WhichMinimizer = WhichMinimizerDefault);
+                            int WhichMinimizer     = WhichMinimizerDefault,
+                            bool UseMultithreading = true);
 
 /**
  * @brief FindNextLocalMinima finds the local minima from the given starting